refactor: one download path for community image endpoints
The gen Worker and the enter registration probe each carried their own
copy of "pull the first image out of an OpenAI images response" — same
base64 branch, same URL branch, same 20 MB cap, drifted timeouts (120s
vs 90s) and two different byte readers.

- Hoists firstCommunityImageBytes() into shared/community-endpoints.ts.
  Callers pass an error factory so gen keeps throwing HttpError(502) and
  enter keeps throwing plain Error for the registration dialog.
- Moves readResponseBytes() to shared/ so the probe gets the streaming
  content-length + running-total cap instead of buffering arrayBuffer().
- Replaces both REQUEST_TIMEOUT_MS constants with one
  COMMUNITY_ENDPOINT_TIMEOUT_MS = 300s, matching the text providers
  (Portkey and Azure both use 290s).
