+++ gen.pollinations.ai/src/userImage.ts
@@ -1,7 +1,7 @@
 import type { ImageInputErrorCode } from "@shared/error.ts";
 import { detectImageMimeType } from "@shared/image-mime.ts";
+import { readResponseBytes } from "@shared/response-bytes.ts";
 import { HttpError } from "./image/httpError.ts";
-import { readResponseBytes } from "./utils/response-bytes.ts";
 import { validateUserMediaUrl } from "./utils/user-media-url.ts";
 
 /**
